pyproject.toml
[project]
name = "wishlist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.127.0",
    "psycopg[binary,pool]>=3.3.2",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "uvicorn>=0.40.0",
    "orjson>=3.11.5",
    "bcrypt==5.0.0",
    "redis>=7.1.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
]
